Add Parts Index and Create

+++ AucklandTalyer/Models/tblParts.cs
@@ -8,7 +8,7 @@ namespace AucklandTalyer.Models
         public int PartsId { get; set; }
         [Required]
         public string PartsName { get; set; }
-        public float? PartsPrice { get; set; }
+        public decimal? PartsPrice { get; set; }
         public DateTime DateAdded { get; set; }
         public string? AddedBy { get; set; }
     }
